requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
altgraph==0.17
async-timeout==4.0.2
atomicwrites==1.4.0
attrs==20.3.0
certifi==2020.12.5
chardet==4.0.0
charset-normalizer==3.0.1
click==8.0.1
colorama==0.4.4
colorgram.py==1.2.0
docutils==0.18.1
ffmpeg-python==0.2.0
Flask==2.0.1
flit==3.7.1
flit_core==3.7.1
frozenlist==1.3.3
future==0.18.2
idna==2.10
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.1
line-bot-sdk==1.19.0
lxml==4.8.0
MarkupSafe==2.0.1
multidict==6.0.4
numpy==1.20.0
openai==0.26.5
opencv-python==4.5.1.48
packaging==20.9
pandas==1.2.4
pefile==2021.5.24
Pillow==8.1.2
pluggy==0.13.1
prettytable==2.1.0
py==1.10.0
pyinstaller==4.3
pyinstaller-hooks-contrib==2021.2
pyparsing==2.4.7
pyperclip==1.8.2
pytest==6.2.2
python-dateutil==2.8.1
pytz==2021.1
pywin32-ctypes==0.2.0
requests==2.25.1
segment==0.2.0
six==1.15.0
toml==0.10.2
tomli==2.0.1
tomli_w==1.0.0
tqdm==4.64.1
twstock @ file:///D:/twstock
urllib3==1.26.4
video-to-ascii==1.2.9
wcwidth==0.2.5
Werkzeug==2.0.1
xtermcolor==1.3
yarl==1.8.2
